+++ rex/cmds/__init__.py
@@ -7,6 +7,7 @@ from rex.version import __version__
 from .emb import emb
 from .new import new
 from .train import train
+from .gen_conf import gen_conf
 
 
 @register("rex_init_call")